fixing type R_init on batched_pythorch_icp

+++ src/bubble_control/bubble_pose_estimation/batched_pytorch_icp.py
@@ -16,7 +16,7 @@ def icp_2d_masked(pc_model, pc_scene, pc_scene_mask, num_iter=30):
         pc_scene_mask = pc_scene_mask.unsqueeze(-1).repeat_interleave(n_coords, dim=-1)  # (N, n_scene_points, n_coords)
     pc_scene_mask_r = reshape_pc(pc_scene_mask)
 
-    R_init = torch.eye(n_coords).unsqueeze(0).repeat_interleave(N, dim=0)  # (N, num_dims, num_dims)--- init all R as identyty
+    R_init = torch.eye(n_coords).unsqueeze(0).repeat_interleave(N, dim=0).type(pc_scene.dtype)  # (N, num_dims, num_dims)--- init all R as identyty
     t_init = masked_tensor_mean(pc_scene_r.transpose(1, 2), pc_scene_mask_r.transpose(1, 2),
                                 start_dim=-1)  # mean of the scene
